pyproject.toml
[project]
name = "financial-analytics"
version = "0.1.0"
description = "FastAPI service with async PDF to Markdown conversion and QA over generated markdowns."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.30.0",
    "langfuse>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.24.0",
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "pyjwt>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "torch; sys_platform == 'linux'",
    "torchvision; sys_platform == 'linux'",
]

[tool.uv]
package = false

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "sys_platform == 'linux'" }]
torchvision = [{ index = "pytorch-cpu", marker = "sys_platform == 'linux'" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
